fix: modal transition
